Improved cBlockHandler::DropBlock

+++ src/Mobs/Monster.cpp
@@ -283,7 +283,7 @@ void cMonster::Tick(float a_Dt, cChunk & a_Chunk)
 			}
 		}
 
-		Vector3f Distance = m_Destination - GetPosition();
+		Vector3d Distance = m_Destination - GetPosition();
 		if (!ReachedDestination() && !ReachedFinalDestination())  // If we haven't reached any sort of destination, move
 		{
 			Distance.y = 0;
